fix(models): RMSNorm weight 转 input_dtype 修复 fp16 Q/K/V dtype 不一致
weight 是 fp32 Parameter，与 fp16 hidden_states 乘会被 promote 到 fp32
导致 xformers memory_efficient_attention 报 Q/K/V dtype 不一致崩溃
（Q/K 过 RMSNorm 后变 fp32，V 保持 fp16）

+++ models/gte-new/modeling.py
@@ -338,7 +338,8 @@ class RMSNorm(nn.Module):
         hidden_states = hidden_states.to(torch.float32)
         variance = hidden_states.pow(2).mean(-1, keepdim=True)
         hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
-        return self.weight * hidden_states.to(input_dtype)
+        # weight 是 fp32 Parameter，与 fp16 乘会 promote 到 fp32，导致下游 xformers 报 Q/K/V dtype 不一致
+        return self.weight.to(input_dtype) * hidden_states.to(input_dtype)
 
 
 LAYER_NORM = {
